+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.2.2] / 2022-08-08
+## [0.3.0] / 2022-08-08
 ### Features
 - Add `IExternalCommandAvailability` on `NewToggleButtonData` Command
 ### Updated
@@ -170,6 +170,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.3.0]: ../../compare/0.2.1...0.3.0
 [0.2.1]: ../../compare/0.2.0...0.2.1
 [0.2.0]: ../../compare/0.1.2...0.2.0
 [0.1.2]: ../../compare/0.1.1...0.1.2